+++ sdk/spring/azure-spring-cloud-stream-binder-servicebus-core/src/main/java/com/azure/spring/servicebus/stream/binder/ServiceBusMessageChannelBinder.java
@@ -115,6 +115,7 @@ public abstract class ServiceBusMessageChannelBinder<T extends ServiceBusExtende
                                      .setMaxConcurrentCalls(consumerProperties.getMaxConcurrentCalls())
                                      .setMaxConcurrentSessions(consumerProperties.getMaxConcurrentSessions())
                                      .setServiceBusReceiveMode(consumerProperties.getServiceBusReceiveMode())
+                                     .setEnableAutoComplete(consumerProperties.isEnableAutoComplete())
                                      .build();
     }